requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.2.0
chardet==3.0.4
discord.py==1.5.1
idna==2.10
requests==2.25.0
beautifulsoup4==4.9.3
multidict==4.7.6
numpy==1.19.3
pandas==1.1.3
Pillow==8.0.1
PyNaCl==1.4.0
python-dateutil==2.8.1
dnspython==2.0.0
asyncio==3.9.0
pytz==2020.1
six==1.15.0
typing-extensions==3.7.4.3
yarl==1.5.1
